Create MeshImporter and move import code from Renderer
